requirements.txt
crawl4ai>=0.5.0
streamlit>=1.30.0
aiohttp>=3.8.4
aiofiles>=23.2.1
playwright>=1.40.0
beautifulsoup4>=4.12.2
faiss-cpu>=1.7.4
ollama>=0.1.5
Pillow>=10.0.0
sentence-transformers>=2.2.2
numpy>=1.24.4
pandas>=2.0.3
tqdm>=4.65.0
python-dotenv>=1.0.0
matplotlib>=3.7.2
mistune>=3.0.1
pygments>=2.15.1